Fix file extension and make some code clean and concise.
