+++ ui/src/hooks/catalog.ts
@@ -1,19 +1,16 @@
 import { useMutation, useQuery, useQueryClient } from '@tanstack/react-query';
 import { CLIENT } from '../context/catalog';
-import { route as _route, isError, assertNever } from '../utils/openapi';
+import { route, isError, assertNever, Router } from '../utils/openapi';
 import type {
   paths as CatalogApi,
   components as CatalogComponent,
 } from '../types/api/catalog.gen';
 import type {
-  Router,
   Model,
   PathParam,
   RequestBody,
   SuccessResponseBody,
 } from '../utils/openapi';
-
-const route: Router<CatalogApi> = _route;
 
 export type CatalogInterface = Model<CatalogComponent, 'CatalogInfo'>;
 
@@ -21,7 +18,7 @@ export function useListCatalogs() {
   return useQuery<SuccessResponseBody<CatalogApi, '/catalogs', 'get'>>({
     queryKey: ['listCatalogs'],
     queryFn: async () => {
-      const api = route({
+      const api = (route as Router<CatalogApi>)({
         client: CLIENT,
         request: {
           path: '/catalogs',
@@ -52,7 +49,7 @@ export function useGetCatalog({ name }: UseGetCatalogArgs) {
   return useQuery<SuccessResponseBody<CatalogApi, '/catalogs/{name}', 'get'>>({
     queryKey: ['getCatalog', name],
     queryFn: async () => {
-      const api = route({
+      const api = (route as Router<CatalogApi>)({
         client: CLIENT,
         request: {
           path: '/catalogs/{name}',
@@ -97,7 +94,7 @@ export function useCreateCatalog() {
       comment,
       properties,
     }: CreateCatalogMutationParams) => {
-      const api = route({
+      const api = (route as Router<CatalogApi>)({
         client: CLIENT,
         request: {
           path: '/catalogs',
@@ -155,7 +152,7 @@ export function useUpdateCatalog({ name }: UseUpdateCatalogArgs) {
       properties,
       new_name,
     }: UpdateCatalogMutationParams) => {
-      const api = route({
+      const api = (route as Router<CatalogApi>)({
         client: CLIENT,
         request: {
           path: '/catalogs/{name}',
@@ -206,7 +203,7 @@ export function useDeleteCatalog() {
     DeleteCatalogMutationParams
   >({
     mutationFn: async ({ name }: DeleteCatalogMutationParams) => {
-      const api = route({
+      const api = (route as Router<CatalogApi>)({
         client: CLIENT,
         request: {
           path: '/catalogs/{name}',
