+++ ui/src/hooks/schemas.ts
@@ -5,21 +5,18 @@ import {
   UseQueryOptions,
 } from '@tanstack/react-query';
 import { CLIENT } from '../context/catalog';
-import { route as _route, isError, assertNever } from '../utils/openapi';
+import { route, isError, assertNever, Router } from '../utils/openapi';
 import type {
   paths as CatalogApi,
   components as CatalogComponent,
 } from '../types/api/catalog.gen';
 import type {
-  Router,
   Model,
   PathParam,
   QueryParam,
   RequestBody,
   SuccessResponseBody,
 } from '../utils/openapi';
-
-const route: Router<CatalogApi> = _route;
 
 export type SchemaInterface = Model<CatalogComponent, 'SchemaInfo'>;
 
@@ -34,7 +31,7 @@ export function useListSchemas({ catalog_name, options }: UseListSchemasArgs) {
   return useQuery<SuccessResponseBody<CatalogApi, '/schemas', 'get'>>({
     queryKey: ['listSchemas', catalog_name],
     queryFn: async () => {
-      const api = route({
+      const api = (route as Router<CatalogApi>)({
         client: CLIENT,
         request: {
           path: '/schemas',
@@ -75,7 +72,7 @@ export function useGetSchema({ full_name }: UseGetSchemaArgs) {
   >({
     queryKey: ['getSchema', catalog, schema],
     queryFn: async () => {
-      const api = route({
+      const api = (route as Router<CatalogApi>)({
         client: CLIENT,
         request: {
           path: '/schemas/{full_name}',
@@ -121,7 +118,7 @@ export function useCreateSchema() {
       comment,
       properties,
     }: CreateSchemaMutationParams) => {
-      const api = route({
+      const api = (route as Router<CatalogApi>)({
         client: CLIENT,
         request: {
           path: '/schemas',
@@ -178,7 +175,7 @@ export function useUpdateSchema({ full_name }: UseUpdateSchemaArgs) {
     UpdateSchemaMutationParams
   >({
     mutationFn: async ({ comment }: UpdateSchemaMutationParams) => {
-      const api = route({
+      const api = (route as Router<CatalogApi>)({
         client: CLIENT,
         request: {
           path: '/schemas/{full_name}',
@@ -235,7 +232,7 @@ export function useDeleteSchema({ full_name }: UseDeleteSchemaArgs) {
     DeleteSchemaMutationParams
   >({
     mutationFn: async ({ full_name }: DeleteSchemaMutationParams) => {
-      const api = route({
+      const api = (route as Router<CatalogApi>)({
         client: CLIENT,
         request: {
           path: '/schemas/{full_name}',
