+++ ui/src/hooks/volumes.ts
@@ -5,21 +5,18 @@ import {
   UseQueryOptions,
 } from '@tanstack/react-query';
 import { CLIENT } from '../context/catalog';
-import { route as _route, isError, assertNever } from '../utils/openapi';
+import { route, isError, assertNever, Router } from '../utils/openapi';
 import type {
   paths as CatalogApi,
   components as CatalogComponent,
 } from '../types/api/catalog.gen';
 import type {
-  Router,
   Model,
   PathParam,
   QueryParam,
   RequestBody,
   SuccessResponseBody,
 } from '../utils/openapi';
-
-const route: Router<CatalogApi> = _route;
 
 export type VolumeInterface = Model<CatalogComponent, 'VolumeInfo'>;
 
@@ -38,7 +35,7 @@ export function useListVolumes({
   return useQuery<SuccessResponseBody<CatalogApi, '/volumes', 'get'>>({
     queryKey: ['listVolumes', catalog_name, schema_name],
     queryFn: async () => {
-      const api = route({
+      const api = (route as Router<CatalogApi>)({
         client: CLIENT,
         request: {
           path: '/volumes',
@@ -74,7 +71,7 @@ export function useGetVolume({ name }: UseGetVolumeArgs) {
   return useQuery<SuccessResponseBody<CatalogApi, '/volumes/{name}', 'get'>>({
     queryKey: ['getVolume', catalog, schema, volume],
     queryFn: async () => {
-      const api = route({
+      const api = (route as Router<CatalogApi>)({
         client: CLIENT,
         request: {
           path: '/volumes/{name}',
@@ -123,7 +120,7 @@ export function useUpdateVolume({ name }: UseUpdateVolumeArgs) {
     UpdateVolumeMutationParams
   >({
     mutationFn: async ({ comment }: UpdateVolumeMutationParams) => {
-      const api = route({
+      const api = (route as Router<CatalogApi>)({
         client: CLIENT,
         request: {
           path: '/volumes/{name}',
@@ -180,7 +177,7 @@ export function useDeleteVolume({ name }: UseDeleteVolumeArgs) {
     DeleteVolumeMutationParams
   >({
     mutationFn: async ({ name }: DeleteVolumeMutationParams) => {
-      const api = route({
+      const api = (route as Router<CatalogApi>)({
         client: CLIENT,
         request: {
           path: '/volumes/{name}',
